use chrome.permissions for fined grain access
Granting an extension access to all possible URLs is not pleasant.
It's also not necessary -- we only need access to the user's VLC
instance.

So drop the mandatory permissions setting and switch to the optional
permissions.  Update the options page to request specific perms when
the user sets things up.

The update process should be seamless as people who have installed
previously have already granted it access to all URLs.